# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel API",
        "description": "API documentation for the Hotel Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/guests": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Get all guests (paginated, filterable by id)",
                "operationId": "56cc8fe33a100159d774722901ae37e8",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Filter by guest id (single or array)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of guests"
                    }
                }
            },
            "post": {
                "tags": [
                    "Guests"
                ],
                "summary": "Create a new guest",
                "operationId": "e792b91966348e6bb3439a3d24883e1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "external_id",
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "external_id": {
                                        "type": "string"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guest created"
                    }
                }
            }
        },
        "/api/guests/{id}": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Get a guest by ID",
                "operationId": "4267610a6a43088364913e1b6071b25d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guest details"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Guests"
                ],
                "summary": "Update a guest",
                "operationId": "af858a767bd4c0c6e7fd978a27775871",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest updated"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guests"
                ],
                "summary": "Delete a guest",
                "operationId": "627107a5c10ad7a3877e6ca545859ca8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Guest deleted"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        },
        "/api/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get all rooms (paginated, filterable by id)",
                "operationId": "db2a0f9c242ba089e87d5ae1fa8e14c5",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Filter by room id (single or array)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of rooms"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create a new room",
                "operationId": "809158a6b73bac8545f93f5b9f8a8a8a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "external_id",
                                    "room_type_id",
                                    "number",
                                    "floor"
                                ],
                                "properties": {
                                    "external_id": {
                                        "type": "string"
                                    },
                                    "room_type_id": {
                                        "type": "integer"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "floor": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created"
                    }
                }
            }
        },
        "/api/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a room by ID",
                "operationId": "b1475eb28ef61d930c8d73dce18cd23c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room details"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update a room",
                "operationId": "0c313fa83b1cc3ffd17495d103f021e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "room_type_id": {
                                        "type": "integer"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "floor": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Room updated"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete a room",
                "operationId": "4698523ee5dafa1ac652ef8ede5e96d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Room deleted"
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        },
...
    },
...
}